% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_fluc}
\alias{check_fluc}
\title{Check TMLE Fluctuation Model}
\usage{
check_fluc(tilt_mod, tilt_tol = 100)
}
\arguments{
\item{tilt_mod}{...}

\item{tilt_tol}{...}
}
\description{
Check TMLE Fluctuation Model
}
