% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nuisance.R
\name{cv_fit_nuisance}
\alias{cv_fit_nuisance}
\title{Fit Nuisance Parameters in a Cross-Validation Fold}
\usage{
cv_fit_nuisance(fold, data, g_stack, e_stack, m_stack, b_stack, d_stack)
}
\arguments{
\item{fold}{...}

\item{data}{...}

\item{g_stack}{...}

\item{e_stack}{...}

\item{m_stack}{...}

\item{b_stack}{...}

\item{d_stack}{...}
}
\description{
Fit Nuisance Parameters in a Cross-Validation Fold
}
