% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimators}
\alias{estimators}
\title{Compute One-step and TML estimators}
\usage{
estimators(
  data,
  delta,
  g_stack,
  e_stack,
  m_stack,
  b_stack,
  d_stack,
  cv_folds = 5L,
  tiltmod_tol = 10
)
}
\arguments{
\item{data}{...}

\item{delta}{...}

\item{g_stack}{...}

\item{e_stack}{...}

\item{m_stack}{...}

\item{b_stack}{...}

\item{d_stack}{...}

\item{cv_folds}{...}

\item{tiltmod_tol}{...}
}
\description{
Compute One-step and TML estimators
}
