% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nuisance.R
\name{fit_nuisance}
\alias{fit_nuisance}
\title{Fit Nuisance Parameters}
\usage{
fit_nuisance(data, g_stack, e_stack, m_stack, b_stack, d_stack, cv_folds = 5L)
}
\arguments{
\item{data}{...}

\item{g_stack}{...}

\item{e_stack}{...}

\item{m_stack}{...}

\item{b_stack}{...}

\item{d_stack}{...}

\item{cv_folds}{...}
}
\description{
Fit Nuisance Parameters
}
