% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gdelta1}
\alias{gdelta1}
\title{Incremental Propensity Score Intervention}
\usage{
gdelta1(g, delta)
}
\arguments{
\item{g}{...}

\item{delta}{...}
}
\description{
Incremental Propensity Score Intervention
}
